refactor(entities): перенес advert-list & clinic-list в ui сегменты соотв-их срезов слоя entities
