+++ src/entities/advert/ui/AdvertCard/index.ts
@@ -1,0 +1,3 @@
+import AdvertCard from './ui';
+
+export { AdvertCard };
