+++ src/entities/advert/ui/AdvertCard/ui/index.tsx
