+++ src/entities/advert/ui/advert-list/index.ts
