+++ src/entities/advert/ui/advert-list/ui/index.scss
