+++ src/entities/clinic/ui/ClinicCard/index.ts
@@ -1,0 +1,3 @@
+import { ClinicCard } from './ui';
+
+export { ClinicCard };
