+++ src/entities/clinic/ui/ClinicCard/ui/index.scss
