--- src/entities/clinic/ui/FullCardClinic/index.tsx
@@ -1,3 +1,0 @@
-export function FullCardClinic() {
-  return null;
-}
