+++ src/entities/clinic/ui/clinic-list/index.ts
