+++ src/entities/clinic/ui/clinic-list/ui/index.scss
