+++ src/entities/clinic/ui/clinic-list/ui/index.tsx
@@ -1,0 +1,13 @@
+import './index.scss';
+
+import { ClinicCard } from '../../ClinicCard';
+
+export default function ClinicList() {
+  return (
+    <div className="clinic-list">
+      <ClinicCard />
+      <ClinicCard />
+      <ClinicCard />
+    </div>
+  );
+}
