+++ src/pages/main/index.ts
@@ -1,0 +1,3 @@
+import MainPage from './ui';
+
+export { MainPage };
