+++ src/pages/main/ui/index.scss
@@ -1,23 +1,16 @@
-@import '../../../shared/lib/styles/variables';
-
-.card-list {
-  width: 350px;
-  background-color: $color-card-background;
-  font-family: $base-font-family;
-  height: 100%;
-  overflow-y: auto;
-}
-
-.card-list::-webkit-scrollbar {
-  display: none;
+.main-page {
+  display: flex;
+  gap: 30px;
+  width: 1110px;
+  margin: 0 auto;
+  padding-bottom: 30px;
 }
 
 @media screen and (max-width: 1027px) {
-  .card-list {
-    min-width: 290px;
-    display: flex;
+  .main-page {
     flex-direction: column;
-    background-color: $color-base-background;
-    order: 1;
+    gap: 40px;
+    align-items: center;
+    max-width: 680px;
   }
 }
