--- src/widgets/clinic-list/ui/index.tsx
@@ -1,13 +1,0 @@
-import './index.scss';
-
-import { SmallCardClinic } from '~/entities/clinic';
-
-export default function ClinicList() {
-  return (
-    <div className="clinic-list">
-      <SmallCardClinic />
-      <SmallCardClinic />
-      <SmallCardClinic />
-    </div>
-  );
-}
